Update 2.0.0 A
BIGGEST UPDATE YET!
Patch notes:
All variable are new configurable (RGB Values, Size of screenshots, filename, etc)!
Can be used universally!
